+++ hiqlite/src/lib.rs
@@ -34,7 +34,7 @@ pub use crate::query::rows::Row;
 #[cfg(feature = "sqlite")]
 pub use crate::store::state_machine::sqlite::{
     param::Param,
-    state_machine::Params,
+    state_machine::{Params, RaftSerializedTimestamp, RaftSerializedTimestampTransaction},
     transaction_variable::{StmtColumn, StmtIndex},
 };
 #[cfg(feature = "dlock")]